Use consistent type of quote for strings
I really did not want to be the one to pick the type of quote and I suspect I'll be lynched whichever type I pick. The decision to use single quotes was made by following the first style guide I thought of consulting, which was JavaScript Standard Style. It may or may not have been subconsciously affected by my time as a Python developer.